Add LTTNG_CONSUMER_CREATE_TRACE_CHUNK command to consumerd protocol
Add a new trace chunk creation command to the sessiond <-> consumerd
protocol. This command will be used to create a trace chunk to
which stream files will be associated on the consumer daemon's end.

The command is implemented by the consumer daemon and used by
the session daemon in follow-up commits.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>