Implement the LTTNG_CONSUMER_CREATE_TRACE_CHUNK consumerd command
The LTTNG_CONSUMER_CREATE_TRACE_CHUNK command creates a chunk on
the consumer daemon's end.

The trace chunk registry, as used by the consumer daemon, implicitly
owns the trace chunks. This is only needed in the consumer since
the consumer has no notion of a session beyond session IDs being
used to identify other objects.

The life time of trace chunks in the consumer daemon is managed by
the session daemon through the LTTNG_CONSUMER_CREATE_TRACE_CHUNK
and LTTNG_CONSUMER_RELEASE_TRACE_CHUNK commands.

The implementation of the LTTNG_CONSUMER_RELEASE_TRACE_CHUNK
is added in a follow-up commit.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>