Add LTTNG_CONSUMER_RELEASE_TRACE_CHUNK command to consumerd protocol
Add a new trace chunk release command to the sessiond <-> consumerd
protocol. This command will be used to release the global reference
to a given trace chunk. Reelasing the consumer daemon's global
reference to the trace chunk leaves only the streams to hold
references until the moment when they are either closed or they
switch-over to another chunk in the event of a session rotation.

The command is implemented by the consumer daemon and used by
the session daemon in follow-up commits.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>