Add LTTNG_CONSUMER_CLOSE_TRACE_CHUNK command to consumerd protocol
Add a new trace chunk close command to the sessiond <-> consumerd
protocol. This command is used to release the global reference to a
given trace chunk in the consumer daemon. Relasing the consumer
daemon's global reference to the trace chunk leaves only the streams
to hold references until the moment when they are either closed or
they switch-over to another chunk in the event of a session rotation.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>