Add Babeltrace2-based logging helpers as liblogging
Add an internal liblogging helper library to be used in the overhaul
of the logging output. The terminal color handling code of babeltrace2
is also added as part of that library.

The babeltrace2 logging is based on the zf_log code, itself MIT
licensed.

This adds the LTTNG_TERM_COLOR environment variable which can be set
to "always" and "never" to override the behaviour resulting from the
auto-detection logic in case of problems.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>