Fix: relayd: protocol error results in NULL dereference
When a session is stopped and rotated, its streams are rotated
out of their current chunk to a "null" chunk. This means that
the streams don't have a configured output until the session
is re-started.

As the session is started, a new trace chunk is announced to the
consumer and relay daemons and that trace chunk is then used as
the output of the existing streams.

However, a protocol error could cause new packets to be sent to
the relay daemon while no output is configured. This would result
in a NULL pointer dereference as the streams attempt to flush a
packet's contents through a non-existant output.

This fix adds a check to ensure to relayd_stream's write() method,
that the stream currently belongs to a trace chunk and has a
"stream_fd". On error, a warning indicating the protocol error is
logged and the connection is torn down.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>