relayd: fail to start a session if it doesn't have a trace chunk
While the relayd start command is unused by recent versions of the
consumer daemon, the semantics of the command are still interesting: a
consumer can ask if the relayd is ready to collect traces for a given
session.

Report that the session doesn't have a current trace chunk, which
would prevent any trace data from being captured for that session in
the current implementation of the relay daemon.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>