Setup trigger group for registering app
Create a pipe for each app and setup a trigger group associated with that
pipe. Transfer the write side to the app, and transfer the read side to
the notification thread as an application event source (*_ADD_APPLICATION)

TODO: what if the app does not support trigger group?
Today, the communication handle will be null on failure to create the
trigger group. The app will be skipped based on this for token update.

Note that we also handle the scenario where the app is killed during the
trigger group creation. The communication handle serve as a
feature check. We might want to have an official flag?

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I3e4aab84e3270ddef1f50f72f946a4d80b3f36e0
Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479